+++ backend/src/main/java/coffeeshout/global/websocket/ui/dto/RecoveryMessage.java
@@ -1,17 +1,19 @@
-package coffeeshout.global.websocket.recovery.dto;
+package coffeeshout.global.websocket.ui.dto;
 
 import coffeeshout.global.websocket.ui.WebSocketResponse;
+import lombok.extern.slf4j.Slf4j;
 
 /**
  * 복구용 메시지 DTO
  *
- * @param messageId Hash 기반 메시지 ID
+ * @param streamId Redis Stream Entry ID (복구 요청 시 lastId로 사용)
  * @param destination 원래 웹소켓 destination (프론트 라우팅용)
  * @param response WebSocketResponse 객체
  * @param timestamp 메시지 생성 시간 (epoch millis)
  */
+@Slf4j
 public record RecoveryMessage(
-        String messageId,
+        String streamId,
         String destination,
         WebSocketResponse<?> response,
         long timestamp
