+++ backend/src/main/java/coffeeshout/global/websocket/ui/dto/RecoveryResponse.java
@@ -1,4 +1,4 @@
-package coffeeshout.global.websocket.recovery.dto;
+package coffeeshout.global.websocket.ui.dto;
 
 import java.util.List;
